:recycle: refactor ProductData, Product: 타입 변경 Int -> Long

+++ app/src/main/java/woowacourse/shopping/data/source/DummyShoppingCartProductIdDataSource.kt
@@ -6,7 +6,7 @@ import woowacourse.shopping.data.model.ProductIdsCountData
 class DummyShoppingCartProductIdDataSource(
     private val pagingStrategy: NumberPagingStrategy<ProductIdsCountData> = NumberPagingStrategy(5),
 ) : ShoppingCartProductIdDataSource {
-    override fun findByProductId(productId: Int): ProductIdsCountData? = productIdsCount.find { it.productId == productId }
+    override fun findByProductId(productId: Long): ProductIdsCountData? = productIdsCount.find { it.productId == productId }
 
     override fun loadAll(): List<ProductIdsCountData> = productIdsCount
 
@@ -14,24 +14,24 @@ class DummyShoppingCartProductIdDataSource(
 
     override fun loadPaged(page: Int): List<ProductIdsCountData> = pagingStrategy.loadPagedData(page, productIdsCount)
 
-    override fun addedNewProductsId(productIdsCountData: ProductIdsCountData): Int {
+    override fun addedNewProductsId(productIdsCountData: ProductIdsCountData): Long {
         productIdsCount.add(productIdsCountData)
         return productIdsCountData.productId
     }
 
-    override fun removedProductsId(productId: Int): Int {
+    override fun removedProductsId(productId: Long): Long {
         val foundItem = productIdsCount.find { it.productId == productId } ?: throw NoSuchElementException()
         productIdsCount.remove(foundItem)
         return foundItem.productId
     }
 
-    override fun plusProductsIdCount(productId: Int) {
+    override fun plusProductsIdCount(productId: Long) {
         val oldItem = productIdsCount.find { it.productId == productId } ?: throw NoSuchElementException()
         productIdsCount.remove(oldItem)
         productIdsCount.add(oldItem.copy(quantity = oldItem.quantity + 1))
     }
 
-    override fun minusProductsIdCount(productId: Int) {
+    override fun minusProductsIdCount(productId: Long) {
         val oldItem = productIdsCount.find { it.productId == productId } ?: throw NoSuchElementException()
         productIdsCount.remove(oldItem)
         productIdsCount.add(oldItem.copy(quantity = oldItem.quantity - 1))
