+++ app/src/main/java/woowacourse/shopping/domain/model/ProductCountEvent.kt
@@ -1,9 +1,9 @@
 package woowacourse.shopping.domain.model
 
 sealed interface ProductCountEvent {
-    data class ProductCountCountChanged(val productId: Int, val count: Int) : ProductCountEvent
+    data class ProductCountCountChanged(val productId: Long, val count: Int) : ProductCountEvent
 
-    data class ProductCountCleared(val productId: Int) : ProductCountEvent
+    data class ProductCountCleared(val productId: Long) : ProductCountEvent
 
     data object ProductCountAllCleared : ProductCountEvent
 }
