+++ app/src/main/java/woowacourse/shopping/domain/repository/ShoppingCartItemRepository.kt
@@ -3,13 +3,13 @@ package woowacourse.shopping.domain.repository
 import woowacourse.shopping.data.model.ProductData
 
 interface ShoppingCartItemRepository {
-    fun addCartItem(product: ProductData): Int
+    fun addCartItem(product: ProductData): Long
 
-    fun findById(id: Int): ProductData
+    fun findById(id: Long): ProductData
 
     fun loadPagedCartItems(page: Int): List<ProductData>
 
-    fun removeCartItem(productId: Int): ProductData
+    fun removeCartItem(productId: Long): ProductData
 
     fun clearAllCartItems()
