+++ app/src/main/java/woowacourse/shopping/ui/OnProductItemClickListener.kt
@@ -1,5 +1,5 @@
 package woowacourse.shopping.ui
 
 interface OnProductItemClickListener {
-    fun onClick(productId: Int)
+    fun onClick(productId: Long)
 }
